Refactor loadPersistedMongoDB
