+++ src/mongo/commands/executeAllMongoCommand.ts
@@ -5,7 +5,7 @@
 
 import { IActionContext } from "@microsoft/vscode-azext-utils";
 import { executeAllCommandsFromActiveEditor } from "../MongoScrapbook";
-import { loadPersistedMongoDB } from "../registerMongoCommands";
+import { loadPersistedMongoDB } from "./connectMongoDatabase";
 
 export async function executeAllMongoCommand(context: IActionContext) {
     await loadPersistedMongoDB();
