+++ src/mongo/commands/executeMongoCommand.ts
@@ -1,7 +1,7 @@
 import { IActionContext } from "@microsoft/vscode-azext-utils";
 import * as vscode from "vscode";
 import { executeCommandFromActiveEditor } from "../MongoScrapbook";
-import { loadPersistedMongoDB } from "../registerMongoCommands";
+import { loadPersistedMongoDB } from "./connectMongoDatabase";
 
 export async function executeMongoCommand(context: IActionContext, position?: vscode.Position) {
     await loadPersistedMongoDB();
